Rework README
Reword the README to include better usage instructions and
acknowledgements of included works.

Usage instructions mention the purpose of this repo, usage of
submodules and targets.

The acknowledgements include references to devkitPro/3ds_portlibs and
mention the copyright holders for most projects. Cases in which sources
are modified are also mentioned. Readers are instructed to read the
license terms of the projects for more detailed information.

Closes #3